pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
discord = "^1.7.3"
rick-roll = "^0.0.3"
requests = "^2.28.1"
pycord = "^0.1.1"
python-dotenv = "^0.21.0"
"cord.py" = "^1.27a0"
py-cord = "^2.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"